requirements.txt
crewai
crewai[tools]
python-dotenv
langchain_google_genai # gemini models